requirements.txt
Django==3.0.7
djangorestframework==3.11.0
psycopg2==2.8.5
flake8==3.8.3
django-filter